exercises/02-advanced-deep-learning/src/features.py
"""Data Augmentation and Preprocessing for Deep Learning

This module provides:
- ImageAugmenter for advanced data augmentation (rotation, flip, color jitter)
- DatasetBuilder for creating PyTorch datasets with augmentation
- Immediate visual feedback showing augmentation effects

Learning objectives:
1. Implement geometric transformations (rotation, flipping, cropping)
2. Apply color transformations (brightness, contrast, saturation)
3. Use Mixup and CutMix for improved generalization
4. Create train/val/test datasets with appropriate augmentation
5. Visualize augmentation effects for debugging

Advanced augmentation techniques:
- Random erasing: randomly mask patches (improves robustness)
- Mixup: blend two images and labels (alpha = 0.2-0.4)
- CutMix: cut and paste patches between images
- AutoAugment: learned augmentation policies
- Normalize with ImageNet statistics (transfer learning)
"""

import logging
from typing import Optional, Tuple, List
from pathlib import Path

import torch
import torch.nn as nn
from torch.utils.data import Dataset, DataLoader, random_split
import torchvision
import torchvision.transforms as T
from torchvision.datasets import CIFAR10, CIFAR100, ImageFolder
from rich.console import Console
from PIL import Image
import numpy as np

logger = logging.getLogger("deeplearning")
console = Console()


class ImageAugmenter:
    """Advanced image augmentation for deep learning.
    
    Provides composition of geometric and photometric transformations
    with immediate visual feedback.
    
    Standard augmentation pipeline:
    1. Random crop with padding
    2. Random horizontal flip
    3. Color jitter (brightness, contrast, saturation, hue)
    4. Random rotation
    5. Normalize with ImageNet mean/std
    """
    
    def __init__(
        self,
        image_size: int = 224,
        random_crop_padding: int = 4,
        horizontal_flip_prob: float = 0.5,
        rotation_degrees: int = 15,
        color_jitter_strength: float = 0.4,
        random_erasing_prob: float = 0.25,
        normalize: bool = True
    ):
        """Initialize augmentation pipeline.
        
        Args:
            image_size: Target image size (224 for ResNet/ViT)
            random_crop_padding: Padding for random crop
            horizontal_flip_prob: Probability of horizontal flip
            rotation_degrees: Maximum rotation angle (±degrees)
            color_jitter_strength: Strength of color jitter (0.0-1.0)
            random_erasing_prob: Probability of random erasing
            normalize: Apply ImageNet normalization
        """
        self.image_size = image_size
        self.random_crop_padding = random_crop_padding
        self.horizontal_flip_prob = horizontal_flip_prob
        self.rotation_degrees = rotation_degrees
        self.color_jitter_strength = color_jitter_strength
        self.random_erasing_prob = random_erasing_prob
        self.normalize = normalize
        
        # ImageNet statistics for normalization
        self.mean = [0.485, 0.456, 0.406]
        self.std = [0.229, 0.224, 0.225]
    
    def get_train_transform(self) -> T.Compose:
        """TODO: Build training augmentation pipeline with geometric and color transforms."""
        # TODO: Your implementation here
        raise NotImplementedError("Implement training augmentation pipeline")
    
    def get_val_transform(self) -> T.Compose:
        """TODO: Build validation/test augmentation pipeline (no random transforms)."""
        # TODO: Your implementation here
        raise NotImplementedError("Implement validation pipeline")


class MixupCutmixTransform:
    """Mixup and CutMix augmentation for improved generalization.
    
    Mixup (Zhang et al., 2017):
    - Blend two images: x_mixed = lambda * x1 + (1 - lambda) * x2
    - Blend labels: y_mixed = lambda * y1 + (1 - lambda) * y2
    - lambda ~ Beta(alpha, alpha), typically alpha=0.2
    
    CutMix (Yun et al., 2019):
    - Cut rectangular patch from x1, paste into x2
    - Mix labels proportional to patch area
    - Encourages model to use full image context
    """
    
    def __init__(self, mixup_alpha: float = 0.2, cutmix_alpha: float = 1.0, prob: float = 0.5):
        """Initialize Mixup/CutMix transform.
        
        Args:
            mixup_alpha: Mixup alpha parameter (higher = more mixing)
            cutmix_alpha: CutMix alpha parameter
            prob: Probability of applying mixup/cutmix
        """
        self.mixup_alpha = mixup_alpha
        self.cutmix_alpha = cutmix_alpha
        self.prob = prob
    
    def __call__(self, batch: Tuple[torch.Tensor, torch.Tensor]) -> Tuple[torch.Tensor, torch.Tensor]:
        """TODO: Apply Mixup or CutMix augmentation to batch."""
        # TODO: Your implementation here
        raise NotImplementedError("Implement Mixup/CutMix")


class DatasetBuilder:
    """Build PyTorch datasets with augmentation for train/val/test.
    
    Supports:
    - CIFAR-10/100
    - ImageFolder (custom datasets)
    - Automatic train/val/test splits
    - Augmentation pipelines
    """
    
    def __init__(
        self,
        dataset_name: str = "cifar10",
        data_dir: str = "./data",
        val_split: float = 0.1,
        num_workers: int = 4
    ):
        """Initialize dataset builder.
        
        Args:
            dataset_name: Dataset name ('cifar10', 'cifar100', 'imagefolder')
            data_dir: Root directory for dataset
            val_split: Validation split ratio (from training set)
            num_workers: Number of workers for data loading
        """
        self.dataset_name = dataset_name.lower()
        self.data_dir = Path(data_dir)
        self.val_split = val_split
        self.num_workers = num_workers
        
        self.augmenter = ImageAugmenter()
    
    def build_datasets(self) -> Tuple[Dataset, Dataset, Dataset]:
        """TODO: Build train/val/test datasets with augmentation."""
        # TODO: Your implementation here
        raise NotImplementedError("Implement dataset building")
    
    def create_dataloaders(
        self,
        train_dataset: Dataset,
        val_dataset: Dataset,
        test_dataset: Dataset,
        batch_size: int = 32
    ) -> Tuple[DataLoader, DataLoader, DataLoader]:
        """TODO: Create DataLoaders for train/val/test."""
        # TODO: Your implementation here
        raise NotImplementedError("Implement DataLoader creation")


# ============================================
# UTILITY FUNCTIONS
# ============================================

def visualize_augmentation(
    dataset: Dataset,
    num_samples: int = 5,
    save_path: Optional[str] = None
) -> None:
    """Visualize augmentation effects on sample images.
    
    Args:
        dataset: Dataset with augmentation
        num_samples: Number of samples to visualize
        save_path: Path to save visualization (optional)
    """
    # TODO: Optional - implement visualization for debugging
    pass


def compute_dataset_statistics(dataset: Dataset) -> Tuple[torch.Tensor, torch.Tensor]:
    """Compute mean and std of dataset for normalization.
    
    Args:
        dataset: Dataset to compute statistics for
    
    Returns:
        Tuple of (mean, std) tensors
    """
    # TODO: Optional - implement statistics computation
    pass
