exercises/02-advanced-deep-learning/tests/test-api.py
"""Tests for Flask API

Validates API endpoints, request handling, and error cases.
"""

import pytest
import json
import io
from PIL import Image
import numpy as np
from unittest.mock import patch, MagicMock


@pytest.fixture
def client():
    """Create Flask test client."""
    from src.api import app
    
    app.config['TESTING'] = True
    
    with app.test_client() as client:
        yield client


@pytest.fixture
def test_image():
    """Create test image file."""
    # Create a simple RGB image
    image = Image.fromarray(
        np.random.randint(0, 255, (640, 640, 3), dtype=np.uint8)
    )
    
    # Save to bytes buffer
    buffer = io.BytesIO()
    image.save(buffer, format='PNG')
    buffer.seek(0)
    
    return buffer


class TestHealthEndpoint:
    """Test suite for /health endpoint."""
    
    def test_health_check(self, client):
        """Test health check endpoint."""
        response = client.get('/health')
        
        assert response.status_code == 200
        
        data = json.loads(response.data)
        assert 'status' in data
        assert data['status'] == 'healthy'
        assert 'model_loaded' in data
        assert 'device' in data


class TestMetricsEndpoint:
    """Test suite for /metrics endpoint."""
    
    def test_metrics_endpoint(self, client):
        """Test Prometheus metrics endpoint."""
        response = client.get('/metrics')
        
        assert response.status_code == 200
        assert b'productioncv' in response.data


class TestDetectEndpoint:
    """Test suite for /detect endpoint."""
    
    @patch('src.api.model')
    def test_detect_with_valid_image(self, mock_model, client, test_image):
        """Test detection with valid image."""
        # Mock model predictions
        mock_model.return_value = [{
            'boxes': [[50, 50, 150, 150]],
            'labels': [1],
            'scores': [0.95]
        }]
        
        data = {
            'image': (test_image, 'test.png'),
            'confidence_threshold': '0.5',
            'max_detections': '100'
        }
        
        response = client.post(
            '/detect',
            data=data,
            content_type='multipart/form-data'
        )
        
        assert response.status_code == 200
        
        result = json.loads(response.data)
        assert result['success'] is True
        assert 'detections' in result
        assert 'inference_time_ms' in result
    
    def test_detect_without_image(self, client):
        """Test detection without image file."""
        response = client.post('/detect', data={})
        
        assert response.status_code == 400
        
        result = json.loads(response.data)
        assert result['success'] is False
        assert 'error' in result
    
    def test_detect_with_invalid_confidence(self, client, test_image):
        """Test detection with invalid confidence threshold."""
        data = {
            'image': (test_image, 'test.png'),
            'confidence_threshold': '1.5'  # Invalid: > 1.0
        }
        
        response = client.post(
            '/detect',
            data=data,
            content_type='multipart/form-data'
        )
        
        assert response.status_code == 400
        
        result = json.loads(response.data)
        assert result['success'] is False
    
    @patch('src.api.config')
    def test_detect_with_large_file(self, mock_config, client):
        """Test detection with file exceeding size limit."""
        # Mock config to set low file size limit
        mock_config.__getitem__.return_value = {'max_upload_size_mb': 0.001}
        
        # Create large image
        large_image = Image.fromarray(
            np.random.randint(0, 255, (2000, 2000, 3), dtype=np.uint8)
        )
        buffer = io.BytesIO()
        large_image.save(buffer, format='PNG')
        buffer.seek(0)
        
        data = {'image': (buffer, 'large.png')}
        
        response = client.post(
            '/detect',
            data=data,
            content_type='multipart/form-data'
        )
        
        # Should return error due to file size
        assert response.status_code in [400, 500]


class TestInfoEndpoint:
    """Test suite for /info endpoint."""
    
    def test_model_info(self, client):
        """Test model info endpoint."""
        response = client.get('/info')
        
        assert response.status_code == 200
        
        data = json.loads(response.data)
        assert 'model_architecture' in data
        assert 'num_classes' in data
        assert 'target_size_mb' in data
        assert 'target_map' in data
        assert 'device' in data
        assert 'compression_enabled' in data


class TestPerformDetection:
    """Test suite for perform_detection function."""
    
    @patch('src.api.model')
    def test_perform_detection(self, mock_model):
        """Test perform_detection function."""
        from src.api import perform_detection
        
        # Mock model output
        mock_predictions = {
            'boxes': [[50, 50, 150, 150], [200, 200, 300, 350]],
            'scores': [0.95, 0.85],
            'labels': [1, 2]
        }
        
        mock_model.return_value = [mock_predictions]
        mock_model.__getitem__.return_value = mock_predictions
        
        # Create test image
        image = np.random.randint(0, 255, (640, 640, 3), dtype=np.uint8)
        
        # Perform detection
        detections, inference_time = perform_detection(
            image,
            confidence_threshold=0.5,
            max_detections=100
        )
        
        assert isinstance(detections, list)
        assert isinstance(inference_time, float)
        assert inference_time > 0
    
    def test_confidence_filtering(self):
        """Test confidence threshold filtering."""
        from src.api import perform_detection
        
        # This would require mocking the model predictions
        # For now, we test the concept
        
        confidence_threshold = 0.8
        
        # In actual implementation, detections below threshold should be filtered
        assert confidence_threshold > 0
        assert confidence_threshold <= 1.0


class TestLabelMapping:
    """Test suite for label name mapping."""
    
    def test_get_label_name(self):
        """Test label name retrieval."""
        from src.api import get_label_name
        
        # Test valid class IDs
        label = get_label_name(0)
        assert label == 'person'
        
        label = get_label_name(2)
        assert label == 'car'
        
        # Test invalid class ID
        label = get_label_name(999)
        assert 'class_' in label
    
    def test_label_consistency(self):
        """Test label name consistency."""
        from src.api import get_label_name
        
        # Same class ID should return same label
        label1 = get_label_name(5)
        label2 = get_label_name(5)
        
        assert label1 == label2
