exercises/02-advanced-deep-learning/requirements.txt
# Core ML Libraries
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
matplotlib>=3.7.0,<4.0.0
seaborn>=0.12.0,<1.0.0
Pillow>=10.0.0,<11.0.0
scikit-learn>=1.3.0,<2.0.0

# Deep Learning Frameworks
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
tensorflow>=2.13.0,<3.0.0

# Computer Vision
opencv-python>=4.8.0,<5.0.0
albumentations>=1.3.0,<2.0.0

# Object Detection & Segmentation
ultralytics>=8.0.0,<9.0.0  # YOLOv8
transformers>=4.30.0,<5.0.0

# Model Optimization & Deployment
onnx>=1.14.0,<2.0.0
onnxruntime>=1.15.0,<2.0.0
tensorflow-model-optimization>=0.7.0,<1.0.0

# Model Tracking & Serving
mlflow>=2.8.0,<3.0.0
flask>=3.0.0,<4.0.0
gunicorn>=21.2.0,<22.0.0
pydantic>=2.4.0,<3.0.0

# Monitoring
prometheus-client>=0.18.0,<1.0.0

# Configuration
pyyaml>=6.0,<7.0.0

# Development & Testing
pytest>=7.4.0,<8.0.0
pytest-cov>=4.1.0,<5.0.0
black>=23.7.0,<24.0.0
flake8>=6.1.0,<7.0.0
pytest-mock>=3.12.0,<4.0.0
